Add files via upload
The covid-19 pandemic enforces the education system to adopt virtual and blended teaching/learning 
modes. In these modes, instructors use special software and platforms to deliver lecture such as Zoom 
and Big Blue Button (BBB). However, it is challenging for instructors to keep track of attendance 
sheets and participation of students during online sessions. In this project you will be designing and 
implementing a python script that investigates and combines attendance and participation information 
of students attending several lectures of a specific course. Your script will process the input 
information and provide a clear and well formatted output files